Redirect user to the projects page if the project they are trying to view does not exist.
Note that, unlike the Rails 4 in Action tutorial, I am using `find_by`
instead of `find`. `find_by` does not raise an error / exception if it
cannot find the record. This means that the set_project method is
slightly different. However, the tests are the same, and they still
pass.